wechat_ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pycryptodome==3.21.0
python-dotenv==1.0.1
cos-python-sdk-v5==1.9.32
